requirements/base.txt
django-braces==1.3.1
django-reversion==1.8.0
django-sendfile
Django==1.6.1
South==0.8.4
